backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# RAG and LLM frameworks
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
llama-index==0.9.15
haystack-ai==2.8.1

# Agent frameworks
crewai==0.1.15
langgraph==0.0.32
autogen-agentchat==0.2.12
agentops==0.1.4

# Vector databases and embeddings
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.6.1
anthropic==0.8.1

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3
redis==5.0.1
motor==3.3.2  # MongoDB async driver

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
tiktoken==0.5.2

# Cloud and deployment
docker==6.1.3
boto3==1.34.0  # AWS
azure-storage-blob==12.19.0  # Azure
google-cloud-storage==2.10.0  # GCP

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

